backend/requirements.txt
Django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
PyMySQL>=1.1
Pillow>=10.0
qrcode[pil]>=7.4
python-decouple>=3.8
django-filter>=23.3
drf-spectacular>=0.26
